apps/azure-rag-setup/raganything-processor/requirements.txt
# RAG-Anything M365 Integration Dependencies
# Core RAG-Anything framework
raganything[all]>=1.2.8

# Azure integration
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# M365 integration
msal>=1.24.0
requests>=2.31.0

# Document processing
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
json5>=0.9.0

# Logging and monitoring
structlog>=23.0.0
